+++ src/pages/Theaters/Theaters.vue
@@ -1,5 +1,6 @@
 <template>
   <div class="card">
+    <VConfirmDialog> </VConfirmDialog>
     <VDialog header="Limitar Lista" :visible.sync="display">
       <h6>Quantidade</h6>
       <input type="text" v-model="limit" />
@@ -20,27 +21,27 @@
         />
       </template>
     </VDialog>
-    <VDialog header="Criar Filme" :visible.sync="displayC">
-      <h6>TheaterId</h6>
-      <input type="text" v-model="Theaters.theaterId" />
-      <h5>Location:</h5>
-      <h6>Street1</h6>
-      <input type="text" v-model="Theaters.location.address.street1" />
-      <h6>Street2</h6>
-      <input type="text" v-model="Theaters.location.address.street2" />
-      <h6>Zipcode</h6>
-      <input type="text" v-model="Theaters.location.address.zipcode" />
-      <h6>State</h6>
-      <input type="text" v-model="Theaters.location.address.state" />
-      <h6>City</h6>
-      <input type="text" v-model="Theaters.location.address.city" />
-      <h5>Geo:</h5>
-      <h6>Type</h6>
-      <input type="text" v-model="Theaters.location.geo.type" />
-      <h6>Latitude</h6>
-      <input type="text" v-model="Theaters.location.geo.coordinates[0]" />
-      <h6>Longitude</h6>
-      <input type="text" v-model="Theaters.location.geo.coordinates[1]" />
+    <VDialog header="Criar Teatro" :visible.sync="displayC">
+      <div class="dialog">
+        <h6>TheaterId</h6>
+        <input type="text" v-model="Theaters.theaterId" />
+        <h6>Street1</h6>
+        <input type="text" v-model="Theaters.location.address.street1" />
+        <h6>Street2</h6>
+        <input type="text" v-model="Theaters.location.address.street2" />
+        <h6>Zipcode</h6>
+        <input type="text" v-model="Theaters.location.address.zipcode" />
+        <h6>State</h6>
+        <input type="text" v-model="Theaters.location.address.state" />
+        <h6>City</h6>
+        <input type="text" v-model="Theaters.location.address.city" />
+        <h6>Type</h6>
+        <input type="text" v-model="Theaters.location.geo.type" />
+        <h6>Latitude</h6>
+        <input type="text" v-model="Theaters.location.geo.coordinates[0]" />
+        <h6>Longitude</h6>
+        <input type="text" v-model="Theaters.location.geo.coordinates[1]" />
+      </div>
       <template #footer>
         <VButton
           label="Cancelar"
@@ -56,27 +57,27 @@
         />
       </template>
     </VDialog>
-    <VDialog header="Atualizar Filme" :visible.sync="displayU">
-      <h6>TheaterId</h6>
-      <input type="text" v-model="Theaters.location.address.theaterId" />
-      <h5>Location:</h5>
-      <h6>Street1</h6>
-      <input type="text" v-model="Theaters.location.address.street1" />
-      <h6>Street2</h6>
-      <input type="text" v-model="Theaters.location.address.street2" />
-      <h6>Zipcode</h6>
-      <input type="text" v-model="Theaters.location.address.zipcode" />
-      <h6>State</h6>
-      <input type="text" v-model="Theaters.location.address.state" />
-      <h6>City</h6>
-      <input type="text" v-model="Theaters.location.address.city" />
-      <h5>Geo:</h5>
-      <h6>Type</h6>
-      <input type="text" v-model="Theaters.location.geo.type" />
-      <h6>Latitude</h6>
-      <input type="text" v-model="Theaters.location.geo.coordinates[0]" />
-      <h6>Longitude</h6>
-      <input type="text" v-model="Theaters.location.geo.coordinates[1]" />
+    <VDialog :header="updateMessage" :visible.sync="displayU">
+      <div class="dialog">
+        <h6>TheaterId</h6>
+        <input type="text" v-model="Theaters.location.address.theaterId" />
+        <h6>Street1</h6>
+        <input type="text" v-model="Theaters.location.address.street1" />
+        <h6>Street2</h6>
+        <input type="text" v-model="Theaters.location.address.street2" />
+        <h6>Zipcode</h6>
+        <input type="text" v-model="Theaters.location.address.zipcode" />
+        <h6>State</h6>
+        <input type="text" v-model="Theaters.location.address.state" />
+        <h6>City</h6>
+        <input type="text" v-model="Theaters.location.address.city" />
+        <h6>Type</h6>
+        <input type="text" v-model="Theaters.location.geo.type" />
+        <h6>Latitude</h6>
+        <input type="text" v-model="Theaters.location.geo.coordinates[0]" />
+        <h6>Longitude</h6>
+        <input type="text" v-model="Theaters.location.geo.coordinates[1]" />
+      </div>
       <template #footer>
         <VButton
           label="Cancelar"
@@ -92,17 +93,17 @@
         />
       </template>
     </VDialog>
-    <div class="mr-4">
+    <div class="mx-2 my-3">
       <template>
         <VButton
-          label="Listar"
+          label="Listar Todos"
           icon="pi pi-list"
           class="p-button-primary mr-2"
           @click="showData"
         />
-         <VButton
+        <VButton
           label="Limitar Listagem"
-          icon="pi pi-list"
+          icon="pi pi-sliders-h"
           class="p-button-info mr-2"
           @click="showPaginateDialog"
         />
@@ -130,7 +131,7 @@
       :value="this.$store.state.Theaters"
       :paginator="true"
       :rows="10"
-      stripedRows
+      showGridlines
       selectionMode="single"
       @row-select="onRowSelect"
     >
@@ -144,27 +145,27 @@
 <script>
 export default {
   methods: {
-     onRowSelect(event){
-      this.id = event.data._id
+    onRowSelect(event) {
+      this.id = event.data._id;
+      this.TheaterName = event.data.theaterId;
+      this.updateMessage = `Atualizar Teatro: ${this.TheaterName}`;
     },
     showData() {
-      this.$store.dispatch(
-        "getTheaters",
-        `Bearer ${this.$store.state.jwtToken}`
-      );
+      this.$store.dispatch("getTheaters", `Bearer ${this.jwt}`);
     },
-     getTheatersPaginate(){
+    getTheatersPaginate() {
       const data = {
-        jwt: `Bearer ${this.$store.state.jwtToken}`,
+        jwt: `Bearer ${this.jwt}`,
         limit: this.limit,
-        skip: this.skip
-      }
+        skip: this.skip,
+      };
       this.$store.dispatch("getTheatersPaginate", data);
+      this.display = false;
     },
-     showPaginateDialog(){
-      this.display = true
+    showPaginateDialog() {
+      this.display = true;
     },
-      hiddenPaginateDialog() {
+    hiddenPaginateDialog() {
       this.display = false;
     },
     showCreateDialog() {
@@ -180,31 +181,50 @@ export default {
       this.displayU = false;
     },
     deleteTheater() {
-      const data = {
-        id: this.id,
-        jwt: `Bearer ${this.$store.state.jwtToken}`,
-      };
-      this.$store.dispatch("deleteTheaters", data);
+      this.$confirm.require({
+        message: `Deseja Remover: ${this.TheaterName}`,
+        header: "Confirmação",
+        icon: "pi pi-exclamation-triangle",
+        acceptLabel: "Deletar",
+        rejectLabel: "Cancelar",
+        accept: () => {
+          const data = {
+            id: this.id,
+            jwt: `Bearer ${this.jwt}`,
+          };
+          this.$store.dispatch("deleteTheaters", data);
+        },
+      });
     },
     createTheater() {
-      this.Theaters.location.geo.coordinates[0] = parseFloat(this.Theaters.location.geo.coordinates[0]);
-      this.Theaters.location.geo.coordinates[1] = parseFloat(this.Theaters.location.geo.coordinates[1]);
+      this.Theaters.location.geo.coordinates[0] = parseFloat(
+        this.Theaters.location.geo.coordinates[0]
+      );
+      this.Theaters.location.geo.coordinates[1] = parseFloat(
+        this.Theaters.location.geo.coordinates[1]
+      );
       const data = {
         Theaters: this.Theaters,
-        jwt: `Bearer ${this.$store.state.jwtToken}`,
+        jwt: `Bearer ${this.jwt}`,
       };
       this.$store.dispatch("createTheaters", data);
+      this.displayC = false;
     },
     updateTheater() {
-      this.Theaters.location.geo.coordinates[0] = parseFloat(this.Theaters.location.geo.coordinates[0]);
-      this.Theaters.location.geo.coordinates[1] = parseFloat(this.Theaters.location.geo.coordinates[1]);
+      this.Theaters.location.geo.coordinates[0] = parseFloat(
+        this.Theaters.location.geo.coordinates[0]
+      );
+      this.Theaters.location.geo.coordinates[1] = parseFloat(
+        this.Theaters.location.geo.coordinates[1]
+      );
       const data = {
         Theaters: this.Theaters,
-        jwt: `Bearer ${this.$store.state.jwtToken}`,
+        jwt: `Bearer ${this.jwt}`,
         id: this.id,
       };
-      console.log(data)
+      console.log(data);
       this.$store.dispatch("updateTheaters", data);
+      this.displayU = false;
     },
   },
   data() {
@@ -212,6 +232,9 @@ export default {
       display: false,
       displayC: false,
       displayU: false,
+      jwt: localStorage.getItem("token"),
+      TheaterName: "",
+      updateMessage: "Nenhum teatro selecionado",
       limit: "",
       skip: "",
       id: "",
@@ -236,5 +259,7 @@ export default {
 };
 </script>
 <style scoped>
+.dialog {
+  column-count: 3;
+}
 </style>
-
